requirements.txt
psycopg2-binary>=2.9.0
pgvector>=0.2.0
sentence-transformers>=2.2.0
huggingface-hub>=0.20.0
groq>=0.4.0
python-dotenv>=1.0.0
datasets>=2.16.0
numpy>=1.24.0
openai
